Fixed a bug in the cleanup system involving index starting at 1.
Added game events for warping and shooting.
Added the first attempt at an input system and component.
Updated meteoroid blueprint and game to fix issues involving the breaking apart of asteroids.
Added drag to space to make the game easier to play.'
Added a warp component and system to handle the warping of the ship.