Changed locations of data to reflect the new locations thanks to the addition of the Asset Manager.
Changed conversion in numberdisplayelement to use the stringconversion functions.
Added android app glue because it has to be there for the moment.
added needed android items (resources, manifest) to the project files folder (because that is where they have to be for now
renamed project files to project-files (android builder can't handle spaces in directories)
updated gitignore to add the temp-files folder (see previous line to see why we need the hyphen)