Fixed bug that caused ui elements to not be able to be made invisible.
Made all ui elements visible by default.
Fixed a bug in NumberDisplayElement's UpdateDisplayedVertices that made all numbers appear to be positive.
Fixed the location of the minus in negative numbers in NumberDisplayElement when prefix zeroes are being used.
Moved all UI creation functions in MeteoroidGame to a helper function.
Added a score display for the second player (its data is still attached to the first player, though)